refactor: LoadLanding과 IntroSection으로 분리
